requirement.txt
Flask
Flask-Cors
Werkzeug
gunicorn
psycopg2-binary